pyproject.toml
[project]
name = "pcd-gem-engine"
version = "0.1.0"
description = "PCD GEM Concierge Workflow Engine — modular GP deck processing pipeline"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "pymupdf>=1.24.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]
web = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
]

[project.scripts]
gem-engine = "run:main"
